pyproject.toml
[project]
name = "sentiment-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "tweepy>=4.16.0",
]
